removed length restriction on migration file price column - main

+++ backend/db/migrations/02-20240515161237-create-spot.js
@@ -56,7 +56,7 @@ module.exports = {
         allowNull: false
       },
       price: {
-        type: Sequelize.INTEGER(5),
+        type: Sequelize.INTEGER,
         allowNull: false
       },
       // previewImage: {
